src/functions/GenerateAuditCertificate/requirements.txt
azure-functions
azure-data-tables>=12.0.0